common: Move ChainConfig builder into a submodule
